Subscriptions
A new OTT streaming service was recently started in Bengaluru .A group of N friends in Bengaluru want to buy OTT subscriptions. We know that 6 people can share one OTT subscription. Also, the cost of one OTT subscription is X rupees. Determine the minimum total cost that the group of N friends will incur so that everyone in the group is able to use OTT.