Best Of Two
Ram and Shyam are playing a game. Each player rolls a regular six faced dice 3 times.The score of a player is the sum of the values of the highest 2 rolls. The player with the highest score wins, and the game ends in a Tie if both players have the same score.
Find the winner of the game or determine whether it is a Tie.