Dinner Preparation
Shyam and Ram are fighting over who makes the dinner, both of them are very lazy so they don't want to make it.
Shyam has a patience level of A, meaning he will only be convinced on the Ath time when Ram asks him to make the dinner.
Ram, on the other hand, has decided to convince Shyam at most B times. If Shyam is not convinced by the Bth time, Ram will proceed to make it himself.
Given the values of A and B, find out who will end up making the dinner.