Sort it
You are given arrays A and B of sizes N and M respectively, such that 1 ≤ Bi ​≤ N. For each 1 ≤ i ≤ M.  Sort the suffix of length Bi in array A in non-decreasing order.
Output the resultant array A after all operations. Note that a suffix of length X denotes the subarray consisting of last X elements of the array.