RahulEren
Rahul is a very big fan of Eren Yeager. In the last season of Attack on Titan, there were N episodes numbered from 1 to N.
Each even indexed episode was A minutes long and each odd indexed episode was B minutes long.
Calculate the total duration (in minutes) of the last season.