Better Deal
There are 2 stores in Switzerland and both sell the same product. The first store sells the product for 100 rupees whereas the second store sells it for 200 rupees.
It is the holiday season and both stores have announced a special discount. The first store is providing a discount of A percent on its product and the second store is providing a discount of B percent on its product.
Gaurav is wondering which store is selling the product at a cheaper price after the discount has been applied. Can you help him identify the better deal?