Single-use Attack
Ram is playing a video game, and is now fighting the final boss. The boss has H health points. Each attack of Ram reduces the health of the boss by X. Ram also has a special attack that can be used at most once, and will decrease the health of the boss by Y. Ram wins when the health of the boss is ≤ 0.
What is the minimum number of attacks needed by Ram to win?